test: update wiki slug expectations
